test(setup): add TestTryAcquireStaleLockCleanup for non-Windows POSIX platforms
